BAlasbaien C:

+++ TPV2/TPV2/src/components/Gun.h
@@ -7,7 +7,7 @@ class Gun: public Component
 private:
 
 	//Music* shotSound;
-	Uint32 lastTimeShot=250;
+	float lastTimeShot=250;
 	Transform* shipData;
 
 public:
